Corrected domain structure and renamed directories Actions renamed as Ports, as they serve as the connection between the Core and the Adapters Ports renamed to Interfaces, as the classes were actual interfaces
