+++ src/routers/v1/user_router.py
@@ -2,16 +2,19 @@ import inject
 from fastapi import APIRouter, HTTPException, status
 from fastapi.encoders import jsonable_encoder
 
-from domain.dependency_injection.user_actions import UserActions
+from domain.ports.database_port import DatabasePort
 
 
 @inject.autoparams()
-def create_user_router(user_actions: UserActions):
+def create_user_router(database_port: DatabasePort):
     router: APIRouter = APIRouter(prefix="/users")
 
     @router.get("/{user_id}")
     def get_user_by_id(user_id: str):
-        user = user_actions.execute_get_by_id(user_id=user_id)
+        try:
+            user = database_port.get_user_by_id(user_id=user_id)
+        except:
+            raise HTTPException(status.HTTP_400_BAD_REQUEST, "user id is not valid")
         if user is None:
             raise HTTPException(status.HTTP_404_NOT_FOUND, "user not found")
         return jsonable_encoder(user.to_dict())
